requirements.txt
absl-py==0.9.0
APScheduler==3.6.3
argon==0.2.0
asgiref==3.2.10
astunparse==1.6.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.13.1
botocore==1.16.1
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
dill==0.2.5
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.41.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-estimators==0.2.1
django-heroku==0.3.1
django-imperavi==0.2.3
django-markdown==0.8.4
django-markdown-editor==1.0.3
django-newsletter==0.9
django-phonenumber-field==5.0.0
django-pwa==1.0.10
django-rest-framework==0.1.0
django-sendgrid-v5==0.8.1
django-storages==1.9.1
django-summernote==0.8.11.6
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
joblib==0.15.1
ldif3==3.1.1
oauthlib==3.1.0
opt-einsum==3.2.1
phonenumbers==8.12.11
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyreadline==2.1
python-card-me==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
surlex==0.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
